Separate shared NDJSON
